Now plays blip repeatedly!
Added Timer and TimerTask to handle playing the sound at specific
intervals.

The period of the metronome is now set by the user input.

Checks for tempo values between 40 and 200 bpm and displays error
message if the value is out of bounds.

Click stops and starts correctly with button clicks.

All strings now taken from resource files, as it should be!